+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/HlsChunkSource.java
@@ -15,6 +15,8 @@
  */
 package androidx.media3.exoplayer.hls;
 
+import static androidx.media3.exoplayer.upstream.LoadErrorHandlingPolicy.FALLBACK_TYPE_LOCATION;
+import static androidx.media3.exoplayer.upstream.LoadErrorHandlingPolicy.FALLBACK_TYPE_TRACK;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static java.lang.Math.max;
@@ -52,6 +54,9 @@ import androidx.media3.exoplayer.trackselection.BaseTrackSelection;
 import androidx.media3.exoplayer.trackselection.ExoTrackSelection;
 import androidx.media3.exoplayer.upstream.CmcdConfiguration;
 import androidx.media3.exoplayer.upstream.CmcdData;
+import androidx.media3.exoplayer.upstream.LoadErrorHandlingPolicy;
+import androidx.media3.exoplayer.upstream.LoadErrorHandlingPolicy.FallbackOptions;
+import androidx.media3.exoplayer.upstream.LoadErrorHandlingPolicy.FallbackSelection;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
 import com.google.common.primitives.Ints;
@@ -701,46 +706,91 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
   }
 
   /**
-   * Attempts to exclude the track associated with the given chunk. Exclusion will fail if the track
-   * is the only non-excluded track in the selection.
-   *
-   * @param chunk The chunk whose load caused the exclusion attempt.
-   * @param exclusionDurationMs The number of milliseconds for which the track selection should be
-   *     excluded.
-   * @return Whether the exclusion succeeded.
-   */
-  public boolean maybeExcludeTrack(Chunk chunk, long exclusionDurationMs) {
-    return trackSelection.excludeTrack(
-        trackSelection.indexOf(trackGroup.indexOf(chunk.trackFormat)), exclusionDurationMs);
-  }
-
-  /**
    * Called when a playlist load encounters an error.
    *
    * @param playlistUrl The {@link Uri} of the playlist whose load encountered an error.
-   * @param exclusionDurationMs The duration for which the playlist should be excluded. Or {@link
-   *     C#TIME_UNSET} if the playlist should not be excluded.
-   * @return Whether the playlist will be excluded from future loads.
+   * @param fallbackSelection The {@link FallbackSelection selected fallback option} to be conducted
+   *     to handle this error, or {@code null} if no fallback is needed.
+   * @return Whether a fallback has been successfully conducted.
    */
-  public boolean onPlaylistError(Uri playlistUrl, long exclusionDurationMs) {
-    int trackGroupIndex = C.INDEX_UNSET;
-    for (int i = 0; i < redundantGroups.length; i++) {
-      if (redundantGroups[i].equals(playlistTracker.getRedundantGroup(playlistUrl))) {
-        trackGroupIndex = i;
-        break;
-      }
-    }
-    if (trackGroupIndex == C.INDEX_UNSET) {
-      return false;
-    }
-    int trackSelectionIndex = trackSelection.indexOf(trackGroupIndex);
-    if (trackSelectionIndex == C.INDEX_UNSET) {
-      return false;
-    }
+  public boolean onPlaylistError(
+      Uri playlistUrl, @Nullable LoadErrorHandlingPolicy.FallbackSelection fallbackSelection) {
     lastPlaylistErrorUrl = playlistUrl;
-    return exclusionDurationMs != C.TIME_UNSET
-        && trackSelection.excludeTrack(trackSelectionIndex, exclusionDurationMs)
-        && playlistTracker.excludeMediaPlaylist(playlistUrl, exclusionDurationMs);
+    if (fallbackSelection == null) {
+      return false;
+    }
+    long exclusionDurationMs = fallbackSelection.exclusionDurationMs;
+    switch (fallbackSelection.type) {
+      case FALLBACK_TYPE_LOCATION:
+        // We exclude the failed playlist url as location fallback.
+        return playlistTracker.excludeMediaPlaylist(playlistUrl, exclusionDurationMs);
+      case FALLBACK_TYPE_TRACK:
+        int trackGroupIndex = C.INDEX_UNSET;
+        for (int i = 0; i < redundantGroups.length; i++) {
+          if (redundantGroups[i].equals(playlistTracker.getRedundantGroup(playlistUrl))) {
+            trackGroupIndex = i;
+            break;
+          }
+        }
+        if (trackGroupIndex == C.INDEX_UNSET) {
+          return false;
+        }
+        int trackSelectionIndex = trackSelection.indexOf(trackGroupIndex);
+        if (trackSelectionIndex == C.INDEX_UNSET) {
+          return false;
+        }
+        // If track fallback is selected, then we should exclude the corresponding track in the
+        // track selection, and exclude the failed playlist url.
+        return trackSelection.excludeTrack(trackSelectionIndex, exclusionDurationMs)
+            && playlistTracker.excludeMediaPlaylist(playlistUrl, exclusionDurationMs);
+      default:
+        // Shouldn't happen.
+        throw new IllegalStateException(
+            "Invalid fallback selection type: " + fallbackSelection.type);
+    }
+  }
+
+  /**
+   * Called when a {@link Chunk} load encounters an error.
+   *
+   * @param chunk The {@link Chunk} whose load encountered an error.
+   * @param fallbackSelection The {@link FallbackSelection selected fallback option} to be conducted
+   *     to handle this error, or {@code null} if no fallback is needed.
+   * @return Whether a fallback has been successfully conducted.
+   */
+  public boolean onChunkError(
+      Chunk chunk, @Nullable LoadErrorHandlingPolicy.FallbackSelection fallbackSelection) {
+    if (fallbackSelection == null) {
+      return false;
+    }
+    long exclusionDurationMs = fallbackSelection.exclusionDurationMs;
+    switch (fallbackSelection.type) {
+      case FALLBACK_TYPE_LOCATION:
+        if (chunk instanceof HlsMediaChunk) {
+          HlsMediaChunk mediaChunk = (HlsMediaChunk) chunk;
+          // We exclude the playlist url as location fallback. Though we might not have seen
+          // the load failure for the playlist, but it's better to avoid loading chunks from
+          // that playlist.
+          return playlistTracker.excludeMediaPlaylist(mediaChunk.playlistUrl, exclusionDurationMs);
+        }
+        return false;
+      case FALLBACK_TYPE_TRACK:
+        int trackGroupIndex = trackGroup.indexOf(chunk.trackFormat);
+        if (trackGroupIndex == C.INDEX_UNSET) {
+          return false;
+        }
+        int trackSelectionIndex = trackSelection.indexOf(trackGroupIndex);
+        if (trackSelectionIndex == C.INDEX_UNSET) {
+          return false;
+        }
+        // If track fallback is selected, then we should exclude the corresponding track in the
+        // track selection.
+        return trackSelection.excludeTrack(trackSelectionIndex, exclusionDurationMs);
+      default:
+        // Shouldn't happen.
+        throw new IllegalStateException(
+            "Invalid fallback selection type: " + fallbackSelection.type);
+    }
   }
 
   /**
@@ -873,6 +923,74 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
       }
     }
     return false;
+  }
+
+  /**
+   * Creates {@link FallbackOptions} for playlist loading error.
+   *
+   * @param playlistUrl The URL of the playlist that encounters the loading error.
+   */
+  public FallbackOptions createFallbackOptions(Uri playlistUrl) {
+    long nowMs = SystemClock.elapsedRealtime();
+    HlsRedundantGroup currentRedundantGroup =
+        checkNotNull(playlistTracker.getRedundantGroup(playlistUrl));
+    // The numberOfLocations is the number of pathways (urls) in the current variant redundant
+    // group.
+    int numberOfLocations = currentRedundantGroup.size();
+    int numberOfExcludedLocations = 0;
+    for (Uri url : currentRedundantGroup.getAllPlaylistUrls()) {
+      if (playlistTracker.isExcluded(url, nowMs)) {
+        // The numberOfExcludedLocations is the number of pathways (urls) that are excluded in the
+        // current variant redundant group.
+        numberOfExcludedLocations++;
+      }
+    }
+    // The numberOfTracks is the number of tracks in the selection, which should also be equal to
+    // the number of redundant groups of this chunk source.
+    int numberOfTracks = trackSelection.length();
+    int numberOfExcludedTracks = 0;
+    ExoTrackSelection trackSelection = getTrackSelection();
+    for (int i = 0; i < redundantGroups.length; i++) {
+      if (trackSelection.isTrackExcluded(i, nowMs)
+          || playlistTracker.isExcluded(redundantGroups[i], nowMs)) {
+        // The numberOfExcludedTracks is the number of redundant groups in this chunk source
+        // whose corresponding track is excluded from the track selection, or the urls from all
+        // locations are excluded.
+        numberOfExcludedTracks++;
+      }
+    }
+    return new FallbackOptions(
+        numberOfLocations, numberOfExcludedLocations, numberOfTracks, numberOfExcludedTracks);
+  }
+
+  /**
+   * Creates {@link FallbackOptions} for {@link Chunk} loading error.
+   *
+   * @param chunk The {@link Chunk} that encounters the loading error.
+   */
+  public FallbackOptions createFallbackOptions(Chunk chunk) {
+    if (chunk instanceof HlsMediaChunk) {
+      // If the failing chunk is an HlsMediaChunk, then we can follow the logic of creating
+      // fallback options for a playlist url.
+      return createFallbackOptions(((HlsMediaChunk) chunk).playlistUrl);
+    }
+    // Otherwise, this chunk is of another type eg. InitializationChunk that may be specific to a
+    // track, then we do the track fallback.
+    long nowMs = SystemClock.elapsedRealtime();
+    ExoTrackSelection trackSelection = getTrackSelection();
+    int numberOfTracks = trackSelection.length();
+    int numberOfExcludedTracks = 0;
+    for (int i = 0; i < numberOfTracks; i++) {
+      if (trackSelection.isTrackExcluded(i, nowMs)
+          || playlistTracker.isExcluded(redundantGroups[i], nowMs)) {
+        numberOfExcludedTracks++;
+      }
+    }
+    return new FallbackOptions(
+        /* numberOfLocations= */ 1,
+        /* numberOfExcludedLocations= */ 0,
+        numberOfTracks,
+        numberOfExcludedTracks);
   }
 
   // Private methods.